+++ packages/orq-rc/src/lib/config.ts
@@ -58,7 +58,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.11.0-rc.76",
+  sdkVersion: "4.11.0-rc.77",
   genVersion: "2.915.0",
-  userAgent: "speakeasy-sdk/typescript 4.11.0-rc.76 2.915.0 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.11.0-rc.77 2.915.0 2.0 @orq-ai/node",
 } as const;